requirements.txt
streamlit>=1.36.0
lxml==5.2.2
requests>=2.32.3
PyYAML>=6.0.1
phonenumbers>=8.13.40
pandas>=2.2.2
